submissions/german_teacher_Agent_V1/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
gradio>=5.32.0
python-dotenv>=1.0.0
pyyaml>=6.0